backend/services/prediction.py
import pandas as pd
import numpy as np
import datetime
from sqlalchemy.orm import Session
from sqlalchemy import text
import os

# Conditional Imports
try:
    from prophet import Prophet
except ImportError:
    Prophet = None

try:
    from xgboost import XGBRegressor
except ImportError:
    XGBRegressor = None

try:
    from sklearn.preprocessing import MinMaxScaler
    from tensorflow.keras.models import Sequential
    from tensorflow.keras.layers import LSTM, Dense
except ImportError:
    MinMaxScaler = None
    Sequential = None
    LSTM = None
    Dense = None

# Suppress warnings
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'

class PredictionService:
    def __init__(self, db: Session):
        self.db = db
        self.look_back = 24
        
    def get_historical_data(self, days=90) -> pd.DataFrame:
        start_time = datetime.datetime.now() - datetime.timedelta(days=days)
        query = text("""
            SELECT timestamp, vehicle_count, occupancy_rate 
            FROM vehicle_metrics 
            WHERE timestamp >= :start_time 
            ORDER BY timestamp ASC
        """)
        try:
            result = self.db.execute(query, {"start_time": start_time})
            rows = result.fetchall()
        except Exception:
            return pd.DataFrame()
            
        if not rows: return pd.DataFrame()
        
        df = pd.DataFrame(rows, columns=['timestamp', 'vehicle_count', 'occupancy_rate'])
        df['timestamp'] = pd.to_datetime(df['timestamp'])
        
        # Aggregate to hourly
        df = df.set_index('timestamp').resample('H').agg({
            'vehicle_count': 'sum',
            'occupancy_rate': 'mean'
        }).reset_index().fillna(0)
        
        return df

    def generate_forecast(self, days_history=60):
        df = self.get_historical_data(days=days_history)
        if len(df) < 48: 
            return self._generate_fallback_forecast()

        # 1. Prophet
        vals_p = np.zeros(24)
        if Prophet:
            try:
                p_df = df[['timestamp', 'vehicle_count']].rename(columns={'timestamp': 'ds', 'vehicle_count': 'y'})
                m_prophet = Prophet(yearly_seasonality=False, daily_seasonality=True)
                m_prophet.fit(p_df)
                future_p = m_prophet.make_future_dataframe(periods=24, freq='H')
                forecast_p = m_prophet.predict(future_p)
                vals_p = forecast_p['yhat'].tail(24).values
            except Exception as e:
                print(f"Prophet error: {e}")

        # 2. XGBoost
        vals_xgb = np.zeros(24)
        last_ts = df['timestamp'].max()
        future_dates = [last_ts + datetime.timedelta(hours=i+1) for i in range(24)]
        
        if XGBRegressor:
            try:
                df['hour'] = df['timestamp'].dt.hour
                df['dow'] = df['timestamp'].dt.dayofweek
                X = df[['hour', 'dow']]
                y = df['vehicle_count']
                
                model_xgb = XGBRegressor(n_estimators=50)
                model_xgb.fit(X, y)

                X_future = pd.DataFrame({
                    'hour': [d.hour for d in future_dates],
                    'dow': [d.dayofweek for d in future_dates]
                })
                vals_xgb = model_xgb.predict(X_future)
            except Exception as e:
                print(f"XGBoost error: {e}")

        # 3. LSTM
        vals_lstm = vals_xgb # Fallback default
        if Sequential and MinMaxScaler:
            try:
                 scaler = MinMaxScaler()
                 scaled_data = scaler.fit_transform(df['vehicle_count'].values.reshape(-1, 1))
                 
                 X_train, y_train = [], []
                 for i in range(len(scaled_data) - self.look_back):
                     X_train.append(scaled_data[i:i+self.look_back, 0])
                     y_train.append(scaled_data[i+self.look_back, 0])
                 
                 X_train, y_train = np.array(X_train), np.array(y_train)
                 X_train = np.reshape(X_train, (X_train.shape[0], X_train.shape[1], 1))
                 
                 model_lstm = Sequential()
                 model_lstm.add(LSTM(32, input_shape=(self.look_back, 1)))
                 model_lstm.add(Dense(1))
                 model_lstm.compile(loss='mse', optimizer='adam')
                 model_lstm.fit(X_train, y_train, epochs=3, verbose=0, batch_size=16)
                 
                 last_seq = scaled_data[-self.look_back:]
                 curr_seq = last_seq.reshape((1, self.look_back, 1))
                 preds = []
                 for _ in range(24):
                     pred = model_lstm.predict(curr_seq, verbose=0)[0][0]
                     preds.append(pred)
                     curr_seq_flat = curr_seq.flatten()
                     curr_seq_flat = np.roll(curr_seq_flat, -1)
                     curr_seq_flat[-1] = pred
                     curr_seq = curr_seq_flat.reshape((1, self.look_back, 1))
                 vals_lstm = scaler.inverse_transform(np.array(preds).reshape(-1, 1)).flatten()
            except Exception as e:
                print(f"LSTM Error: {e}")

        # Ensemble
        # Handle if models failed or missed
        components = []
        if np.any(vals_p): components.append(vals_p)
        if np.any(vals_xgb): components.append(vals_xgb)
        # vals_lstm logic above defaults to vals_xgb, so need care.
        # Simplify: just average what we have.
        
        if not components:
             return self._generate_fallback_forecast()
             
        ensemble = np.mean(components, axis=0)
        ensemble = np.maximum(ensemble, 0) 
        
        return {
            "forecast_24h": np.round(ensemble, 1).tolist(),
            "peak_demand_time": future_dates[np.argmax(ensemble)].strftime("%H:%M"),
            "peak_demand_value": float(np.max(ensemble)),
            "method": "Ensemble (Available Models)"
        }

    def _generate_fallback_forecast(self):
        return {
            "forecast_24h": [10.0] * 24, # Flatline fallback
            "peak_demand_time": "N/A",
            "peak_demand_value": 10.0,
            "note": "Insufficient data for ML training."
        }
